Allow unsubscribing from event listeners
When we have a connection that persists between several components,
we need to be able to unsubscribe from a single event listener
to avoid errors. If we don't unsubscribe, the component that added
the event listener can have been destroyed, and if so, it will
throw an error because the callback in our component has been
unsubscribed, but the event listener tried to trigger
the callback anyway.

HubConnection in SignalR provides the `off` method for this
purpose, but it's not exposed in ng2-signalr.
We also need to keep track of the callback functions, as SignalR
only allows us to unsubscribe from a single callback. Ideally we
would support a single in ng2-signalr as well, but for now, it
is enough to unsubscribe from all at once.